fix: use NUMERIC instead of BIGINT for uint64 historyId comparison
Gmail historyId is uint64 (max ~18.4e18), but BIGINT is signed (max ~9.2e18).
Using NUMERIC ensures we can handle all valid Gmail history IDs.

Also fix test to mock $executeRaw instead of emailAccount.update.

+++ apps/web/app/api/google/webhook/process-history.ts
@@ -248,7 +248,7 @@ async function updateLastSyncedHistoryId({
     WHERE id = ${emailAccountId}
     AND (
       "lastSyncedHistoryId" IS NULL
-      OR CAST("lastSyncedHistoryId" AS BIGINT) < CAST(${lastSyncedHistoryId} AS BIGINT)
+      OR CAST("lastSyncedHistoryId" AS NUMERIC) < CAST(${lastSyncedHistoryId} AS NUMERIC)
     )
   `;
 }
